backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
onnx
onnxruntime-gpu
pandas
pillow
requests
scikit-learn
torch
torchvision
tqdm
